+++ Source/Modules/Tests/Diagnostics/Gapotchenko.FX.Diagnostics.CommandLine.Tests/CommandLineTests.cs
@@ -1,5 +1,3 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
 namespace Gapotchenko.FX.Diagnostics.CommandLine.Tests;
 
 [TestClass]